Move GothamService into `service` module
