+++ src/clone_sockets.rs
@@ -7,7 +7,8 @@ use tokio_core;
 use tokio_core::reactor::Core;
 use futures::{Future, Stream};
 
-use handler::{NewHandler, GothamService};
+use handler::NewHandler;
+use service::GothamService;
 
 /// Starts a Gotham application, with the given number of threads.
 pub fn start_with_num_threads<NH, A>(addr: A, threads: usize, new_handler: NH)
